Initial stab at dealing with unparsable local package versions
Some AUR package use non-standard version number formats which we can't
be expected to parse. In these cases, we should just compare the local
version string with that of the remote package, and if they are
different, assume that the local package is out of date.

Testing this change was foiled by the fact that the AUR site no longer
supports HTTP (see #3).

This should address #1, once tested.