Omit VCS packages from output
The recorded version number of a VCS package is bumped each time the
package is rebuilt. In most cases, this occurs more frequently than the
remote AUR PKGBUILD is updated. There is no reliable way to tell whether
a locally installed VCS package is truly out of date relative to its
remote version or if merely its version number is out of sync, so just
omit these packages entirely from the output.